refector : step3 자동차경주 피드백 후 수정사항 반영
