+++ src/main/java/step3_racingcar/domain/Car.java
@@ -20,7 +20,4 @@ public class Car {
 		return position;
 	}
 
-	public CarDTO toDTO() {
-		return new CarDTO(position);
-	}
 }
