+++ src/main/java/step3_racingcar/dto/CarDTO.java
@@ -1,10 +1,12 @@
 package step3_racingcar.dto;
 
+import step3_racingcar.domain.Car;
+
 public class CarDTO {
 	private final int position;
 
-	public CarDTO(int position) {
-		this.position = position;
+	public CarDTO(Car car) {
+		this.position = car.getPosition();
 	}
 
 	public int getPosition() {
