+++ src/main/java/step3_racingcar/view/ResultView.java
@@ -2,16 +2,14 @@ package step3_racingcar.view;
 
 import java.util.List;
 
-import step3_racingcar.domain.Car;
 import step3_racingcar.dto.CarDTO;
-import step3_racingcar.service.CarStatus;
 
 public class ResultView {
 
 	private static void printCarStatus(List<CarDTO> carDTOs) {
 		System.out.println();
 		for (CarDTO carDTO : carDTOs) {
-			System.out.println(CarStatus.formatStatus(carDTO.getPosition()));
+			System.out.println(formatStatus(carDTO.getPosition()));
 		}
 	}
 
@@ -19,5 +17,13 @@ public class ResultView {
 		for (List<CarDTO> roundResult : results) {
 			printCarStatus(roundResult);
 		}
+	}
+
+	public static String formatStatus(int position) {
+		StringBuilder status = new StringBuilder();
+		for (int i = 0; i < position; i++) {
+			status.append("-");
+		}
+		return status.toString();
 	}
 }
