+++ src/test/java/step3_racingcar/CarStatusTest.java
@@ -2,10 +2,12 @@ package step3_racingcar;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import javax.xml.transform.Result;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import step3_racingcar.service.CarStatus;
+import step3_racingcar.view.ResultView;
 
 public class CarStatusTest {
 	
@@ -14,7 +16,7 @@ public class CarStatusTest {
 	void formatStatus() {
 		int position = 5;
 		String expectedStatus = "-----";
-		String actualStatus = CarStatus.formatStatus(position);
+		String actualStatus = ResultView.formatStatus(position);
 		assertEquals(expectedStatus, actualStatus);
 	}
 }
